+++ packages/stylelint-config-wordpress/scss.js
@@ -15,6 +15,24 @@ module.exports = {
     "at-rule-no-unknown": [ true, {
       ignoreAtRules: [ "extend", "at-root", "warn", "error", "if", "else", "for", "each", "while", "mixin", "include", "content", "return", "function" ],
     } ],
+    "at-rule-empty-line-before": [
+      "always", {
+        "ignoreAtRules": ["else"],
+      },
+    ],
+    "block-opening-brace-space-before": "always",
+    "block-closing-brace-newline-after": [
+      "always", {
+        "ignoreAtRules": [ "if", "else" ],
+      },
+    ],
+    "at-rule-name-space-after": "always",
+    "rule-non-nested-empty-line-before": "always",
+    "scss/at-else-closing-brace-newline-after": "always-last-in-chain",
+    "scss/at-else-closing-brace-space-after": "always-intermediate",
+    "scss/at-else-empty-line-before": "never",
+    "scss/at-if-closing-brace-newline-after": "always-last-in-chain",
+    "scss/at-if-closing-brace-space-after": "always-intermediate",
     "scss/selector-no-redundant-nesting-selector": true,
   },
 }